ComposeBox: Stop using UNSAFE_componentWillReceiveProps.
This replaces UNSAFE_componentWillReceiveProps with a key attribute, so
that we create a new ComposeBox instance when we or start or stop
editing a message.

This fixes the "Warning: Cannot update a component from inside the
function body of a different component" error that we were getting
several hundred times when a user starts editing a message, although we
don't fully understand why it fixes this error.

The `initialMessage`, `initialTopic`, and `autoFocusMessage` props were
added in this commit to emulate the imperative behaviour in
UNSAFE_componentWillReceiveProps.

See the CZO thread for more info and discussion:
https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Next.20steps.20for.20.60ComposeBox.60/near/1160847