ComposeBox: Send "stop typing" notification to destination narrow.
This sends the "stop typing" notification to the destination narrow,
rather than the narrow that the ComposeBox happened to be loaded into.

This doesn't matter right now, but could in the future when we start
sending typing status notifications to streams.

This isn't totally correct - we may still miss sending some stop typing
notifications that we should send, if a user makes edits to the message
and then later edits the topic, but this change is still an improvement.