notifications: Add more detail to "realm_uri ambiguous" error.
There are a lot of users seeing this - more than I would expect to have
multiple accounts on the same server. I want to collect this data so
that we can have a more complete picture of what's going on in these
cases.

This is somewhat personal data (URLs of self-hosted realms), but we log
this same array elsewhere, so it's not changing the status quo
significantly to log it here. I've purposefully avoided logging email
addresses, instead logging the total number of unique identities. If we
need more information to debug this further, we can add that later.